MAGE-1153 Refactor command classes for extending namespace
